trigger sigs creation from block subround
