+++ consensus/spos/bls/v2/subroundSignature_test.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	"sync"
-	"sync/atomic"
 	"testing"
 
 	"github.com/multiversx/mx-chain-core-go/core"
@@ -1004,142 +1003,6 @@ func TestSubroundSignature_DoSignatureJobForManagedKeys(t *testing.T) {
 		cancel()
 		r := srSignature.DoSignatureJobForManagedKeys(ctx)
 		assert.False(t, r)
-	})
-}
-
-func TestSubroundSignature_CreateSignaturesForManagedKeys(t *testing.T) {
-	t.Parallel()
-
-	t.Run("should work", func(t *testing.T) {
-		t.Parallel()
-
-		container := consensusMocks.InitConsensusCore()
-		enableEpochsHandler := &enableEpochsHandlerMock.EnableEpochsHandlerStub{
-			IsFlagEnabledInEpochCalled: func(flag core.EnableEpochFlag, epoch uint32) bool {
-				return flag == common.AndromedaFlag
-			},
-		}
-		container.SetEnableEpochsHandler(enableEpochsHandler)
-
-		numMultiKeysSignaturesCreated := int32(0)
-
-		signingHandler := &consensusMocks.SigningHandlerStub{
-			CreateSignatureShareForPublicKeyCalled: func(msg []byte, index uint16, epoch uint32, publicKeyBytes []byte) ([]byte, error) {
-				atomic.AddInt32(&numMultiKeysSignaturesCreated, 1)
-				return []byte("SIG"), nil
-			},
-		}
-		container.SetSigningHandler(signingHandler)
-		consensusState := initializers.InitConsensusStateWithKeysHandler(
-			&testscommon.KeysHandlerStub{
-				IsKeyManagedByCurrentNodeCalled: func(pkBytes []byte) bool {
-					return true
-				},
-			},
-		)
-		ch := make(chan bool, 1)
-
-		sr, _ := spos.NewSubround(
-			bls.SrBlock,
-			bls.SrSignature,
-			bls.SrEndRound,
-			roundTimeDuration,
-			0.7,
-			0.85,
-			"(SIGNATURE)",
-			consensusState,
-			ch,
-			executeStoredMessages,
-			container,
-			chainID,
-			currentPid,
-			&statusHandler.AppStatusHandlerStub{},
-		)
-
-		srSignature, _ := v2.NewSubroundSignature(
-			sr,
-			&statusHandler.AppStatusHandlerStub{},
-			&testscommon.SentSignatureTrackerStub{},
-			&consensusMocks.SposWorkerMock{},
-			&dataRetrieverMock.ThrottlerStub{},
-		)
-
-		sr.SetHeader(&block.Header{})
-		sr.SetSelfPubKey("OTHER")
-
-		r := srSignature.CreateSignaturesForManagedKeys(context.TODO())
-		assert.True(t, r)
-
-		assert.Equal(t, int32(9), atomic.LoadInt32(&numMultiKeysSignaturesCreated)) // there are 9 keys in default consensus group config
-	})
-
-	t.Run("should fail", func(t *testing.T) {
-		t.Parallel()
-
-		container := consensusMocks.InitConsensusCore()
-		enableEpochsHandler := &enableEpochsHandlerMock.EnableEpochsHandlerStub{
-			IsFlagEnabledInEpochCalled: func(flag core.EnableEpochFlag, epoch uint32) bool {
-				return flag == common.AndromedaFlag
-			},
-		}
-		container.SetEnableEpochsHandler(enableEpochsHandler)
-
-		numMultiKeysSignaturesCreated := int32(0)
-
-		signingHandler := &consensusMocks.SigningHandlerStub{
-			CreateSignatureShareForPublicKeyCalled: func(msg []byte, index uint16, epoch uint32, publicKeyBytes []byte) ([]byte, error) {
-				atomic.AddInt32(&numMultiKeysSignaturesCreated, 1)
-				return []byte("SIG"), nil
-			},
-		}
-		container.SetSigningHandler(signingHandler)
-		consensusState := initializers.InitConsensusStateWithKeysHandler(
-			&testscommon.KeysHandlerStub{
-				IsKeyManagedByCurrentNodeCalled: func(pkBytes []byte) bool {
-					return true
-				},
-			},
-		)
-		ch := make(chan bool, 1)
-
-		sr, _ := spos.NewSubround(
-			bls.SrBlock,
-			bls.SrSignature,
-			bls.SrEndRound,
-			roundTimeDuration,
-			0.7,
-			0.85,
-			"(SIGNATURE)",
-			consensusState,
-			ch,
-			executeStoredMessages,
-			container,
-			chainID,
-			currentPid,
-			&statusHandler.AppStatusHandlerStub{},
-		)
-
-		srSignature, _ := v2.NewSubroundSignature(
-			sr,
-			&statusHandler.AppStatusHandlerStub{},
-			&testscommon.SentSignatureTrackerStub{},
-			&consensusMocks.SposWorkerMock{},
-			&dataRetrieverMock.ThrottlerStub{
-				CanProcessCalled: func() bool {
-					return false
-				},
-			},
-		)
-
-		sr.SetHeader(&block.Header{})
-		sr.SetSelfPubKey("OTHER")
-
-		ctx, cancel := context.WithCancel(context.TODO())
-		cancel()
-		r := srSignature.CreateSignaturesForManagedKeys(ctx)
-		assert.False(t, r)
-
-		assert.Equal(t, int32(0), atomic.LoadInt32(&numMultiKeysSignaturesCreated))
 	})
 }
